+++ R/write_omv.R
@@ -3,7 +3,12 @@
 #'
 #' @param dtaFrm Data frame to be exported (default: NULL)
 #' @param fleNme Name / position of the output file to be generated ("FILENAME.omv"; default: "")
-#' @return a list containing the meta data (mtaDta, written to metadata.json in the OMV-file), the extended data (xtdDta, written to xdata.json in the OMV-file) and the original data frame (dtaFrm)
+#' @param retDbg Whether to return a list with debugging information (see Value; default: FALSE)
+#' @return a list (if retDbg == TRUE), containing the meta data (mtaDta, metadata.json in the OMV-file), the extended data (xtdDta, xdata.json in the OMV-file) and the original data frame (dtaFrm)
+#'
+#' @details
+#' jamovi has a specific measurement level / type "ID" (in addition to the "standard" ones "Nominal", "Ordinal", and "Continuous"). "ID" is used for columns that contain some form of ID (e.g., a
+#' participant code). In order to set a variable of your data frame to "ID", you have to manually set an attribute \code{jmv-id} (e.g., \code{attr(dtaFrm$column, "jmv-id") = TRUE}).
 #'
 #' @examples
 #' \dontrun{
@@ -16,7 +21,7 @@
 #' # the print-function is only used to force devtools::run_examples() to show the output
 #' # the same applies to the other "print"-functions further down in the code
 #' # on the command line, one can just run: names(wrtDta)
-#' # → "mtaDta" "xtdDta" "dtaFrm"
+#' # -> "mtaDta" "xtdDta" "dtaFrm"
 #' # returns a list with the metadata (mtaDta, e.g., column and data type),
 #' # the extended data (xtdDta, e.g., variable lables), and the data frame (dtaFrm)
 #' # the purpose of these variables is merely for checking (understanding the file format)
@@ -25,15 +30,15 @@
 #' # check whether the file was written to the disk, get the file information (size, etc.)
 #' # and delete the file afterwards
 #' print(list.files(".", "Trial.omv"));
-#' # → "Trial.omv"
+#' # -> "Trial.omv"
 #' print(file.info("Trial.omv")$size);
-#' # → 2111 (size may differ on different OSes)
+#' # -> 2111 (size may differ on different OSes)
 #' unlink("Trial.omv");
 #' }
 #'
 #' @export write_omv
 
-write_omv <- function(dtaFrm = NULL, fleNme = "") {
+write_omv <- function(dtaFrm = NULL, fleNme = "", retDbg = FALSE) {
 
     # check whether dtaFrm is a data frame
     if (is.null(dtaFrm) || ! is.data.frame(dtaFrm) || any(dim(dtaFrm) < 1)) {
@@ -43,143 +48,146 @@ write_omv <- function(dtaFrm = NULL, fleNme = "") {
     if (length(fleNme) <= 0 || ! grepl(".omv", fleNme) || ! dir.exists(dirname(fleNme))) {
         stop(sprintf("Output file name (%s) doesn't have the correct format (e.g., wrong extension) or destination directory doesn't exist.", fleNme));
     }
+    fleNme <- file.path(normalizePath(dirname(fleNme)), fleNme);
 
-    colNum <- dim(dtaFrm)[2]
+    colNum <- dim(dtaFrm)[2];
 
     # initialize metadata.json
-    mtaDta <- list(dataSet = list(rowCount = dim(dtaFrm)[1], columnCount = colNum, removedRows = list(), addedRows = list(), fields = list(), transforms = list()));
-    mtaDta$dataSet$fields <- rep(list(list(name = "", id = NA, columnType = "Data", dataType = "Integer", measureType = "Nominal", formula = "", formulaMessage = "", parentId = 0,
-                                           width = 100, type = "number", importName = "", description = "", transform = 0, edits = list(), missingValues = list(), trimLevels = TRUE,
-                                           filterNo = NA, active = FALSE)), colNum);
+    mtaDta <- mtaGlb;
+    # use the attributes stored in the data frame (for the whole data set) to
+    # update the metadata
+    mtaDta <- setAtt(names(mtaDta), dtaFrm, mtaDta);
+    # the number of rows and columns has to be adjusted to the current data set
+    mtaDta$rowCount    <- dim(dtaFrm)[1];
+    mtaDta$columnCount <- colNum;
+    # create the entries for storing the column specific information
+    mtaDta$fields      <- rep(list(mtaFld), colNum);
 
     # initialize xdata.json
     xtdDta <- list();
 
     # create data.bin
-    binHdl <- file(description = "data.bin", open = "wb");
+    binHdl <- file(description = file.path(tempdir(), "data.bin"),    open = "wb");
 
     # create strings.bin
-    strHdl <- file(description = "strings.bin", open = "wb");
+    strHdl <- file(description = file.path(tempdir(), "strings.bin"), open = "wb");
     strPos <- 0
 
     for (i in seq_len(colNum)) {
-        # assign fields stored in data frame (if the columns contain attributes);
-        for (attNme in names(mtaDta$dataSet$fields[[i]])) {
-            if (! is.null(attr(dtaFrm[[i]], attNme))) {
-                mtaDta$dataSet$fields[[i]][[attNme]] <- attr(dtaFrm[[i]], attNme);
-            }
-        }
-        rm("attNme");
+        # assign attributes that are stored in data frame for this data column
+        # (if available)
+        mtaDta$fields[[i]] <- setAtt(names(mtaDta$fields[[i]]), dtaFrm[[i]], mtaDta$fields[[i]]);
 
         # name
-        mtaDta$dataSet$fields[[i]][["name"]] <- names(dtaFrm[i]);
+        mtaDta$fields[[i]][["name"]] <- names(dtaFrm[i]);
 
         # variable label: if column contains a "jmv-desc" use this, otherwise try whether the data frame contains "variable.labels" and use that
-        if        (! is.null(attr(dtaFrm[[i]], "jmv-desc"))) {
-            mtaDta$dataSet$fields[[i]][["description"]] <- attr(dtaFrm[[i]], "jmv-desc");
-        } else if (! is.null(attr(dtaFrm, "variable.labels"))) {
-            mtaDta$dataSet$fields[[i]][["description"]] <- attr(dtaFrm, "variable.labels")[[names(dtaFrm[i])]];
-        }
-
-        # assess measurementType (if not stored as attribute)
-        if (is.null(attr(dtaFrm[[i]], "measureType"))) {
-            # if the jmv-id marker is set in the original data
-            if (! is.null(attr(dtaFrm[[i]], "jmv-id")) && attr(dtaFrm[[i]], "jmv-id") == TRUE) {
-                mtaDta$dataSet$fields[[i]][["measureType"]] <- "ID";
-            # if original data are text
-            } else if (is.character(dtaFrm[[i]])) {
-                mtaDta$dataSet$fields[[i]][["measureType"]] <- "Nominal";
-            # if original data are a factor
-            } else if (is.factor(dtaFrm[[i]])) {
-                numLvl <- suppressWarnings(as.integer(dtaFrm[[i]]));
-                if (any(is.na(numLvl)) || (stats::sd(diff(numLvl)) > diff(range(numLvl)) / 10)) {
-                    mtaDta$dataSet$fields[[i]][["measureType"]] <- "Nominal";
-                } else {
-                    mtaDta$dataSet$fields[[i]][["measureType"]] <- "Ordinal";
-                }
-                rm("numLvl");
-            # if original data are numeric
-            } else if (is.numeric(dtaFrm[[i]])) {
-                if (all(is.na(dtaFrm[[i]]))) {
-                    if        (is.integer(dtaFrm[[i]])) {
-                        mtaDta$dataSet$fields[[i]][["measureType"]] <- "Nominal";
-                    } else if (is.double(dtaFrm[[i]]))  {
-                        mtaDta$dataSet$fields[[i]][["measureType"]] <- "Continuous";
-                    } else {
-                        stop("Needs to be implemented: All values NA and column type neither integer nor number / double");
-                    }
-                } else {
-                    if ((length(unique(dtaFrm[[i]])) < diff(range(dtaFrm[[i]])) / 5) || (stats::sd(dtaFrm[[i]]) < diff(range(dtaFrm[[i]])) / 10)) {
-                        mtaDta$dataSet$fields[[i]][["measureType"]] <- "Nominal";
-                    } else {
-                        mtaDta$dataSet$fields[[i]][["measureType"]] <- "Continuous";
-                    }
-                }
-            }
+        if        (chkAtt(dtaFrm[[i]], "jmv-desc")) {
+            mtaDta$fields[[i]][["description"]] <- attr(dtaFrm[[i]], "jmv-desc");
+        } else if (chkAtt(dtaFrm, "variable.labels")) {
+            mtaDta$fields[[i]][["description"]] <- attr(dtaFrm, "variable.labels")[[names(dtaFrm[i])]];
         }
 
         # value labels - R-foreign-style
-        if (! is.null(attr(dtaFrm, "label.table")) && ! is.null(attr(dtaFrm, "label.table")[[names(dtaFrm[i])]])) {
-            stop("R-foreign-style value labels need to be implemented.");
+        if (chkAtt(dtaFrm, "label.table")) {
+            stop("R-foreign-style value labels need to be implemented. Please send the data file that caused this problem to sebastian.jentschke@uib.no");
         }
 
+        # assign column from the original data frame to colCrr (so that modifications don't affect the original)
         colCrr <- dtaFrm[[i]]
-        if (is.null(attr(dtaFrm[[i]], "dataType"))) {
-            if (mtaDta$dataSet$fields[[i]][["measureType"]] != "ID") {
-                if (is.character(colCrr)) {
-                    colCrr <- factor(colCrr);
-                    mtaDta$dataSet$fields[[i]][["dataType"]] <- "Text";
-                }
 
-                if (is.factor(colCrr)) {
-                    facLvl <- attr(colCrr, "levels");
-                    facVal <- attr(colCrr, "values");
-                    if (sum(is.na(suppressWarnings(as.numeric(facLvl)))) > 1) {
-                        mtaDta$dataSet$fields[[i]][["dataType"]] <- "Text";
-                    }
-                    if (is.null(facVal)) {
-                        colCrr <- as.integer(colCrr);
-                        facVal <- unique(sort(colCrr));
-                    } else {
-                        colCrr <-  facVal[as.integer(colCrr)]
-                    }
-                    if (length(facVal) > 0) {
-                        xtdDta[[names(dtaFrm[i])]] <- list(labels = lapply(seq_along(facVal), function(i) list(facVal[[i]], facLvl[[i]], facLvl[[i]])));
-                    }
-                    rm("facLvl", "facVal");
-                } else {
-                    mtaDta$dataSet$fields[[i]][["trimLevels"]] <- NULL
-                }
-
-                if (mtaDta$dataSet$fields[[i]][["type"]] == "number") {
-                    if (mtaDta$dataSet$fields[[i]][["dataType"]] != "Decimal") {
-                        if (all(abs(colCrr - round(colCrr)) < sqrt(.Machine$double.eps), na.rm = TRUE)) {
-                            mtaDta$dataSet$fields[[i]][["type"]]     <- "integer";
-                            mtaDta$dataSet$fields[[i]][["dataType"]] <- "Integer";
-                        } else {
-                            mtaDta$dataSet$fields[[i]][["type"]]     <- "number";
-                            mtaDta$dataSet$fields[[i]][["dataType"]] <- "Decimal";
-                        }
-                    }
+        # ID variables represent a special case and are therefore treated first
+        # only if the jmv-id marker is set or if the measureType is set to "ID" in the original data
+        if (chkAtt(dtaFrm[[i]], "jmv-id", TRUE) || chkAtt(dtaFrm[[i]], "measureType", "ID")) {
+            if (is.character(colCrr)) {
+                mtaDta$fields[[i]][["dataType"]] <- "Text";
+                mtaDta$fields[[i]][["type"]]     <- "string";
+            } else if (is.integer(colCrr)) {
+                mtaDta$fields[[i]][["dataType"]] <- "Integer";
+                mtaDta$fields[[i]][["type"]]     <- "integer";
+            }
+            mtaDta$fields[[i]][["measureType"]]  <- "ID";
+        # afterwards, the different variable types for each column of the original data frame are tested
+        # [a] logical
+        } else if (is.logical(colCrr)) {
+            colCrr <- as.integer(colCrr);
+            mtaDta$fields[[i]][["dataType"]]    <- "Integer";
+            mtaDta$fields[[i]][["type"]]        <- "integer";
+            # measureType not set as the correct type ("Nominal") is already the default
+            xtdDta[[names(dtaFrm[i])]] <- list(labels = lapply(0:1, function(i) list(i, as.character(i), as.character(i), FALSE)));
+        # [b] factors
+        } else if (is.factor(colCrr)) {
+            facLvl <- attr(colCrr, "levels");
+            facVal <- attr(colCrr, "values");
+            if (is.null(facVal)) {
+                colCrr <- as.integer(colCrr);
+                facVal <- unique(sort(colCrr));
+                mtaDta$fields[[i]][["dataType"]] <- ifelse(all(!is.na(suppressWarnings(as.integer(facLvl)))) && all(as.character(as.integer(facLvl)) == facLvl), "Integer", "Text")
+            } else {
+                colCrr <- facVal[as.integer(colCrr)]
+                mtaDta$fields[[i]][["dataType"]] <- "Integer";
+            }
+            mtaDta$fields[[i]][["type"]]         <- "integer";
+            # if "measureType" is already stored in the data frame, keep it, otherwise set it to "Ordinal" if the properties indicate it to be likely ("Nominal" is already the default)
+            if (chkFld(mtaDta$fields[[i]], "dataType", "Integer") && length(facVal) > 5 && !any(is.na(c(facVal, colCrr))) && stats::sd(diff(facVal)) < diff(range(colCrr)) / 10) {
+                mtaDta$fields[[i]][["measureType"]] <- ifelse(chkAtt(dtaFrm[[i]], "measureType"), attr(dtaFrm[[i]], "measureType"), "Ordinal");
+            }
+            if (length(facVal) > 0) {
+                xtdDta[[names(dtaFrm[i])]] <- list(labels = lapply(seq_along(facVal), function(i) list(facVal[[i]], facLvl[[i]], facLvl[[i]], FALSE)));
+            }
+            rm(facLvl, facVal);
+        # [c] characters / strings
+        } else if (is.character(colCrr)) {
+            facLvl <- unique(colCrr);
+            facVal <- seq(1, length(facVal));
+            colCrr <- as.integer(as.factor(colCrr));
+            mtaDta$fields[[i]][["type"]]        <- "integer";
+            mtaDta$fields[[i]][["dataType"]]    <- "Text";
+            # measureType not set as the correct type ("Nominal") is already the default
+            if (length(facVal) > 0) {
+                xtdDta[[names(dtaFrm[i])]] <- list(labels = lapply(seq_along(facVal), function(i) list(facVal[[i]], facLvl[[i]], facLvl[[i]], FALSE)));
+            }
+        # [d] numerical (integer / decimals)facLvl <- unique(colCrr)
+        } else if (is.numeric(colCrr)) {
+            if (all(abs(colCrr - round(colCrr)) < sqrt(.Machine$double.eps), na.rm = TRUE)) {
+                colCrr <- as.integer(colCrr);
+                mtaDta$fields[[i]][["type"]]     <- "integer";
+                mtaDta$fields[[i]][["dataType"]] <- "Integer";
+                # if "measureType" is already stored in the data frame, keep it, otherwise assign "Continuous" if there is a high enough value range and variability (sd)
+                if (length(unique(colCrr)) > diff(range(colCrr, na.rm = TRUE)) / 5 && stats::sd(colCrr, na.rm = TRUE) > diff(range(colCrr, na.rm = TRUE)) / 10) {
+                    mtaDta$fields[[i]][["measureType"]] <- ifelse(chkAtt(dtaFrm[[i]], "measureType"), attr(dtaFrm[[i]], "measureType"), "Continuous");
                 }
             } else {
-                if (is.character(colCrr)) {
-                    mtaDta$dataSet$fields[[i]][["type"]]     <- "string";
-                    mtaDta$dataSet$fields[[i]][["dataType"]] <- "Text";
-                } else if (is.integer(colCrr)) {
-                    mtaDta$dataSet$fields[[i]][["type"]]     <- "integer";
-                    mtaDta$dataSet$fields[[i]][["dataType"]] <- "Integer";
-                }
+                mtaDta$fields[[i]][["type"]]     <- "number";
+                mtaDta$fields[[i]][["dataType"]] <- "Decimal";
+                # if "measureType" is already stored in the data frame, keep it, otherwise assign "Continuous"
+                mtaDta$fields[[i]][["measureType"]] <- ifelse(chkAtt(dtaFrm[[i]], "measureType"), attr(dtaFrm[[i]], "measureType"), "Continuous");
             }
+        # [e] dates / times - not implemented yet
+        } else if (all(sapply(colCrr, inherits, c("Date", "POSIXt")))) {
+            stop("Needs to be implemented: Date / Time. Please send the data file that caused this problem to sebastian.jentschke@uib.no");
+        } else {
+            stop(sprintf("Variable type %s not implemented. Please send the data file that caused this problem to sebastian.jentschke@uib.no", class(colCrr)));
         }
 
-        # columnType-specific arguments
-        if (mtaDta$dataSet$fields[[i]][["columnType"]] != "Filter") {
-            mtaDta$dataSet$fields[[i]][["filterNo"]]   <- NULL;
-            mtaDta$dataSet$fields[[i]][["active"]]     <- NULL;
+        # remove atrributes that are only used with specific columnTypes
+        if (all(class(dtaFrm[[i]]) != c("logical", "factor"))) {
+            mtaDta$fields[[i]][["trimLevels"]]               <- NULL;
         }
-        if (is.null(attr(dtaFrm[[i]], "trimLevels")) && mtaDta$dataSet$fields[[i]][["dataType"]] != "Text") {
-            mtaDta$dataSet$fields[[i]][["trimLevels"]] <- NULL;
+        if (! chkFld(mtaDta$fields[[i]], "columnType", "Filter")) {
+            mtaDta$fields[[i]][["filterNo"]]                 <- NULL;
+            mtaDta$fields[[i]][["active"]]                   <- NULL;
+        } else {
+            # if the variable is a filter, trimLevels is removed even in cases
+            # where the original variable was a factor / logical
+            mtaDta$fields[[i]][["trimLevels"]]               <- NULL;
+        }
+        if (! chkFld(mtaDta$fields[[i]], "columnType", "Output")) {
+            mtaDta$fields[[i]][["outputAnalysisId"]]         <- NULL;
+            mtaDta$fields[[i]][["outputOptionName"]]         <- NULL;
+            mtaDta$fields[[i]][["outputName"]]               <- NULL;
+            mtaDta$fields[[i]][["outputDesiredColumnName"]]  <- NULL;
+            mtaDta$fields[[i]][["outputAssignedColumnName"]] <- NULL;
         }
 
         # fix problem with transforms
@@ -188,14 +196,14 @@ write_omv <- function(dtaFrm = NULL, fleNme = "") {
         # check that dataType, and measureType are set accordingly to type (attribute and column in the data frame)
         # dataType: Text, Integer, Decimal
         #
-        # print(sprintf("%02d: %s - %s - %s\n", i, mtaDta$dataSet$fields[[i]][["type"]], mtaDta$dataSet$fields[[i]][["dataType"]], mtaDta$dataSet$fields[[i]][["measureType"]]));
+        # cat(sprintf("%02d: %s - %s - %s\n", i, mtaDta$fields[[i]][["type"]], mtaDta$fields[[i]][["dataType"]], mtaDta$fields[[i]][["measureType"]]));
 
         # write to data.bin according to type
-        if      (mtaDta$dataSet$fields[[i]][["type"]] == "integer") {
+        if        (chkFld(mtaDta$fields[[i]], "type", "integer")) {
             colWrt <- as.integer(colCrr);
-        } else if (mtaDta$dataSet$fields[[i]][["type"]] == "number") {
+        } else if (chkFld(mtaDta$fields[[i]], "type", "number"))  {
             colWrt <- as.double(colCrr);
-        } else if (mtaDta$dataSet$fields[[i]][["type"]] == "string") {
+        } else if (chkFld(mtaDta$fields[[i]], "type", "string"))  {
             colWrt <- rep(0, length(colCrr));
             for (j in seq_along(colCrr)) {
                 writeBin(colCrr[j], strHdl);
@@ -204,109 +212,144 @@ write_omv <- function(dtaFrm = NULL, fleNme = "") {
             }
             colWrt <- as.integer(colWrt);
         } else {
-            stop(sprintf("Variable type %s not implemented.", mtaDta$dataSet$fields[[i]][["type"]]));
+            stop(sprintf("Variable type %s not implemented. Please send the data file that caused this problem to sebastian.jentschke@uib.no", mtaDta$fields[[i]][["type"]]));
         }
         writeBin(colWrt, binHdl, endian = "little");
 
-        # check that dataType, and measureType are set accordingly
-        rm("colCrr", "colWrt");
-    }
-
-    # attach the attributes stored in the data frame
-    for (attNme in names(mtaDta$dataSet)) {
-        if (! is.null(attr(dtaFrm, attNme))) {
-            mtaDta$dataSet[[attNme]] <- attr(dtaFrm, attNme);
-        }
+        # remove temporary variables for modifying and storing the current column from the data set
+        rm(colCrr, colWrt);
     }
 
     # double check whether ID is unique
-    id_Lst <- unlist(lapply(seq_along(mtaDta$dataSet$fields), function(i) mtaDta$dataSet$fields[[i]][["id"]]));
+    id_Lst <- unlist(lapply(seq_along(mtaDta$fields), function(i) mtaDta$fields[[i]][["id"]]));
     if (any(is.na(id_Lst)) || any(duplicated(id_Lst))) {
-        for (i in seq_along(mtaDta$dataSet$fields)) {
-            mtaDta$dataSet$fields[[i]][["id"]] <- i;
+        for (i in seq_along(mtaDta$fields)) {
+            mtaDta$fields[[i]][["id"]] <- i;
         }
     }
 
     # compress data.bin and discard the temporary file
-    close(binHdl);
-    utils::zip(fleNme, "data.bin", flags = "-r9Xq");
-    unlink("data.bin");
+    add2ZIP(fleNme, binHdl, newFle = TRUE);
+    rm(binHdl);
 
-    # check whether data were written to strings.bin
-    close(strHdl);
-    if (strPos > 0) utils::zip(fleNme, "strings.bin", flags = "-r9Xq");
-    unlink("strings.bin");
+    # compress strings.bin (only if it contains data) and discard the temporary file
+    add2ZIP(fleNme, strHdl, blnZIP = strPos > 0);
+    rm(strHdl);
 
     # create meta, write it and add it to ZIP file
-    # this needs a bit of special handling because of \r\n on Windows vs. \n on Mac / Linux
-    # thanks to MAgojam for figuring out a solution
-    mnfTxt <- enc2utf8(c("Manifest-Version: 1.0",
-                         "Data-Archive-Version: 1.0.2",
-                         "jamovi-Archive-Version: 8.0",
-                         paste("Created-By: jmvReadWrite", utils::packageVersion("jmvReadWrite"))));
-
-    writeLines(mnfTxt, mnfHdl <- file("meta", open = "wb"), sep = "\n"); close(mnfHdl);
-    utils::zip(fleNme, "meta", flags = "-r9Xq");
-    unlink("meta");
-    rm("mnfTxt", "mnfHdl");
+    mnfHdl <- file(file.path(tempdir(), "meta"),         open = "wb");
+    add2ZIP(fleNme, mnfHdl, txtOut = mnfTxt());
+    rm(mnfHdl);
 
     # write metadata.json
-    writeLines(gsub(":", ": ", gsub(",", ", ", rjson::toJSON(mtaDta))), "metadata.json");
-    utils::zip(fleNme, "metadata.json", flags = "-r9Xq");
-    unlink("metadata.json");
+    mtaHdl <- file(file.path(tempdir(), "metadata.json"), open = "w");
+    add2ZIP(fleNme, mtaHdl, txtOut = fmtJSON(list(dataSet = mtaDta)));
+    rm(mtaHdl);
 
     # write xdata.json
-    writeLines(gsub(":", ": ", gsub(",", ", ", rjson::toJSON(xtdDta))), "xdata.json");
-    utils::zip(fleNme, "xdata.json", flags = "-r9Xq");
-    unlink("xdata.json");
+    xtdHdl <- file(file.path(tempdir(), "xdata.json"),    open = "w");
+    add2ZIP(fleNme, xtdHdl, txtOut = fmtJSON(xtdDta));
+    rm(xtdHdl);
 
-    # export empty HTML results output - index.html
-    # TO-DO: implement properly writing HTML if this is implemented in read_omv
-    resHTM <- c("<!DOCTYPE html>",
-                "<html>",
-                "    <head>",
-                "        <meta charset=\"utf-8\" />",
-                "        <title>Results</title>",
-                "        <style>\n",
-                "            body {",
-                "                font-family: \"Segoe UI\",Roboto,Helvetica,Arial,sans-serif,\"Segoe UI Emoji\",\"Segoe UI Symbol\" ;",
-                "                font-size: 12px ;",
-                "                color: #333333 ;",
-                "                margin: 24px;",
-                "                cursor: default ;",
-                "            }\n",
-                "            h1 {",
-                "                font-size: 160% ;",
-                "                color: #3E6DA9 ;",
-                "                margin-bottom: 12px ;",
-                "                white-space: nowrap ;",
-                "            }\n",
-                "            h2 {",
-                "                font-size: 130% ;",
-                "                color: #3E6DA9 ;",
-                "                margin-bottom: 12px ;",
-                "            }\n",
-                "            h3, h4, h5 {",
-                "                font-size: 110% ;",
-                "                margin-bottom: 12px ;",
-                "            }\n",
-                "            table {",
-                "                page-break-inside: avoid;",
-                "                border-spacing: 0 ;",
-                "            }\n",
-                "            table tr td, table tr th {",
-                "                font-size: 12px ;",
-                "                page-break-inside: avoid;",
-                "            }\n",
-                "        </style>",
-                "</head>",
-                "<body>\n",
-                "</body>",
-                "</html>");
-    # write HTML and add it to ZIP file
-    writeLines(resHTM, con = "index.html");
-    utils::zip(fleNme, "index.html", flags = "-r9Xq");
-    unlink("index.html");
+    # write index.html and add it to ZIP file
+    # currently, the HTML that is stored in the HTML attribute can't be saved because it is only a "front"
+    # that doesn't work without the analyses and images contained in it being stored too
+    htmHdl <- file(file.path(tempdir(), "index.html"),    open = "w");
+#   if (!is.null(attr(dtaFrm, "HTML"))) {
+#      add2ZIP(fleNme, htmHdl, txtOut = attr(dtaFrm, "HTML"));
+#   } else {
+#      add2ZIP(fleNme, htmHdl, txtOut = htmTxt());
+#   }
+    add2ZIP(fleNme, htmHdl, txtOut = htmTxt());
+    rm(htmHdl);
 
-    list(mtaDta = mtaDta, xtdDta = xtdDta, dtaFrm = dtaFrm);
+    if (retDbg) {
+        list(mtaDta = mtaDta, xtdDta = xtdDta, dtaFrm = dtaFrm)
+    }
+}
+
+fmtJSON <- function(txtJSON = "") {
+    gsub("  ", " ", gsub(":", ": ", gsub(",", ", ", rjson::toJSON(txtJSON))))
+}
+
+htmTxt <- function() {
+    c("<!DOCTYPE html>",
+      "<html>",
+      "    <head>",
+      "        <meta charset=\"utf-8\" />",
+      "        <title>Results</title>",
+      "        <style>\n",
+      "            body {",
+      "                font-family: \"Segoe UI\",Roboto,Helvetica,Arial,sans-serif,\"Segoe UI Emoji\",\"Segoe UI Symbol\" ;",
+      "                font-size: 12px ;",
+      "                color: #333333 ;",
+      "                margin: 24px;",
+      "                cursor: default ;",
+      "            }\n",
+      "            h1 {",
+      "                font-size: 160% ;",
+      "                color: #3E6DA9 ;",
+      "                margin-bottom: 12px ;",
+      "                white-space: nowrap ;",
+      "            }\n",
+      "            h2 {",
+      "                font-size: 130% ;",
+      "                color: #3E6DA9 ;",
+      "                margin-bottom: 12px ;",
+      "            }\n",
+      "            h3, h4, h5 {",
+      "                font-size: 110% ;",
+      "                margin-bottom: 12px ;",
+      "            }\n",
+      "            table {",
+      "                page-break-inside: avoid;",
+      "                border-spacing: 0 ;",
+      "            }\n",
+      "            table tr td, table tr th {",
+      "                font-size: 12px ;",
+      "                page-break-inside: avoid;",
+      "            }\n",
+      "        </style>",
+      "</head>",
+      "<body>\n",
+      "</body>",
+      "</html>")
+}
+
+mnfTxt <- function() {
+    crrTxt <- paste0(sapply(lstMnf, "[[", 1), ":");
+    for (i in seq_along(lstMnf)) {
+        if (crrTxt[i] == "Created-By:") {
+            crrTxt[i] <- paste(crrTxt[i], "jmvReadWrite", utils::packageVersion("jmvReadWrite"));
+        } else {
+            crrTxt[i] <- paste(crrTxt[i], paste(lstMnf[[i]][2:length(lstMnf[[i]])], collapse = "."));
+        }
+    }
+    enc2utf8(crrTxt)
+}
+
+add2ZIP <- function(fleZIP = "", crrHdl = NULL, newFle = FALSE, blnZIP = TRUE, txtOut = "") {
+
+    if (! all(class(crrHdl) == c("file", "connection"))) {
+        cat(utils::str(crrHdl));
+        stop("Parameter isn't the file handle pointing to a file to be zipped.")
+    }
+
+    if (all(nchar(txtOut) > 0)) {
+        # writing the manifest requires a bit of special handling because of \r\n on Windows vs. \n on Mac / Linux
+        # thanks to MAgojam for figuring out a solution
+        # for the other files (metadata.json, xdata.json, index.html), it doesn't matter, therefore sep = "\n" is added for all
+        writeLines(txtOut, crrHdl, sep = "\n");
+    }
+    crrFle <- summary(crrHdl)$description;
+    close(crrHdl);
+
+    if (blnZIP) {
+        if (newFle) {
+            zip::zip(fleZIP,        basename(crrFle), root = dirname(crrFle));
+        } else {
+            zip::zip_append(fleZIP, basename(crrFle), root = dirname(crrFle));
+        }
+    }
+    unlink(crrFle); rm(crrFle);
 }